requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
dotenv==0.9.9
frozenlist==1.7.0
idna==3.10
multidict==6.6.4
propcache==0.3.2
pyaes==1.6.1
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.1.1
pytun-pmd3==3.0.2
PyYAML==6.0.2
TgCrypto==1.2.5
yarl==1.20.1
